bedrock, corn and symbiotic

+++ utils/adapter.ts
@@ -32,7 +32,7 @@ const runAdapter = async (adapter: AdapterExport, address: string) => {
   ret.points = convertValuesToNormal(ret.points);
   ret.total =
     typeof ret.total === "number"
-      ? parseInt(String(ret.total)) || 0
+      ? parseFloat(String(ret.total)) || 0
       : convertValuesToInt(ret.total);
   ret.claimable = Boolean(ret.claimable);
